Reflect Container.content{Width,Height} through implicitContent{Width,Height}
When a Container does not have an explicit content size set, we pick up
the content size from the implicit content size, so the two are in sync.

However, if an explicit content size is set, we were failing to reflect
this through the control's implicit content size, which means that any
client of Container can not treat it as a plain Control, reading its
implicit content size to do layouting.

We now keep the two values in sync, even when an explicit content size
has been set.

Change-Id: If1253a1de804d99232b6c78ec9eae88f3937efe0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 0234ec9ef17bb4ffccde993d194baed6efae3f63)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>