Reflect Pane.content{Width,Height} through implicitContent{Width,Height}
When a Pane does not have an explicit content size set, we pick up the
content size from the implicit content size, so the two are in sync.

However, if an explicit content size is set, we were failing to reflect
this through the control's implicit content size, which means that any
client of Pane can not treat it as a plain Control, reading its implicit
content size to do layouting.

We now keep the two values in sync, even when an explicit content size
has been set.

As a drive-by this also fixed a bug in QQuickPane::resetContentWidth()
where we would unset hasContentHeight instead of hasContentWidth.

Change-Id: If24b15bef51d3841932d2bf6d645087ccca677a9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 7c968ac1512b8eede38f213708f314701547c784)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>