QmlCompiler: Name types when initializing structured value types
The conversion is not required to produce the same type. It can also
produce something equivalent that holds the same value. For example an
integer of a different kind that on assignment gets C++-converted.

This also exposes that the read type is not necessarily expressible in
C++. To make it expressible, merge it with the type present in the
register.

Task-number: QTBUG-132345
Change-Id: I07accb38062b7c89ab60d50bf6cbd63423e68301
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 70d9a5cbcd1ef482e4f192201a2c60a742607e6b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 18f9246d2e7648318008b4fada4cc6155153e0ad)