Enable read-backs for DateObject
Certain objects in V4 are `ReferenceObject`s, keeping tracks of the data
they originate from and performing read/write-backs from/to it to keep
the data in sync between the reference object and the original data.

For example, some C++ types are converted to reference objects, so that
mutations on the QML side and the C++ side are represented on both
sides.

`DateObject`, which implementes the Javascript `Date` object, is a
`ReferenceObject` and is used, for example, to track `QTime`, `QDate`
and `QDateTime` objects coming from the C++ side.

While it has implemented the required infrastructure for read-backs for
quite a while, it has not really made use of it in its methods, so that
they would generally never consider the original data even whey they are
supposed to.

Thus, ensure that a `DateObject` performs a read-back in its various
"getter"/"conversion" methods, when appropriate, to remain in sync with
its origin.

Add a series of test cases to ensure that the affected methods will now
read-back as required.

Change-Id: I48e841ff0f176f3ab6ab08ede999352f71297fca
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>