Fix propagation of QContextMenuEvent for QQuickItems
QQuickDeliveryAgentPrivate::deliverContextMenuEvent() relies on that the
QContextMenuEvent gets ignored if the target item doesn't handle it,
therefore we have to make sure it is ignored by default.

This fixes the context menu for the SideBar (still WIP) and likewise the
autotest in this change.

Change-Id: If72a814adc92aea6e5e071eb49cbe84b025768fb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 04d75342dd57b033036484f212b7b6a1577b6530)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>