Check if showView uses updated position
ShowView sets a position to a qquickview. The position takes
some time to be updated. In some cases a race condition can
happen on Opensuse and the un-updated position is used.

Check if the qquickview has a frame. If it does not have
one, return true. If it has a frame, verify that its
position is correct.

Check if the starting framePosition and ending position of
the qquickview are the same. Use qWaitFor to check if the
position updated. Return false if it haven't.

If the qquickview does not have a frame, the position check
is not reached.

Fixes: QTBUG-133341
Fixes: QTBUG-132648
Fixes: QTBUG-132630
Fixes: QTBUG-133342
Fixes: QTBUG-133343
Fixes: QTBUG-132941
Fixes: QTBUG-133344
Pick-to: 6.9 6.8 6.5
Task-number: QTBUG-133340
Change-Id: I866a9d2bd2499cafcbfa4ccd53a770846590a7bd
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>