VectorImage: Fix issue with animation ending early
If a transform animation ends while others are still running,
we need to record the state directly after the shorter animation
ends, so that we get a key frame for only the animations that
are still active as a starting point.

Note: This also adds some extra debug output which was helpful
when debugging it.

Fixes: QTBUG-132408
Change-Id: Ied1f4e7431d201f7bcef27eb4be271eb0afbd8f2
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit c85d635d5f0188c42e03fb604c770ebaecc7d761)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>