qmlformat: add custom comment anchoring for case
Add a way to customize where comments are anchored. Prior to this
commit, comments could only be anchored to an AST::Node
firstSourceLocation() or lastSourceLocation().

This commit extends the current functionality to be able to anchor
comments somewhere between firstSourceLocation() and
lastSourceLocation().

This allows to fix the bug where comments would be printed before the
":" in case statements, because they would get attached to the
expression inside the case-statement, instead of getting attached to
the colon of the case-statement.

Extend AstRangesVisitor to add the sourcelocation of the ":" in case-
statements to the list of sourcelocations where comments can be
attached to.

The anchors enum is not meant to be extended: instead, an anchor is
either DefaultAnchor for comments attached to AST::Node*, or
token.begin() for some QQmlJS::SourceLocation token inside an
AST::Node*.

Pick-to: 6.5
Fixes: QTBUG-132886
Change-Id: I0a18f09b3631b8798b6697298e532b47ddc7cb20
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit b91ca3e0bc0b7e0b1ec53b41698d2325337dae59)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d5b0705f4dbf6a25c375ecdb82e28f2ca8123bc5)