QtQml: Store script CUs in QQmlMetaTypeData
When importing a JavaScript module again from the same QML engine, the
resulting executable compilation unit should be the same as before. This
only works if we use the same base compilation unit. However, so far
QQmlScriptBlob did not store the compilation unit in the central type
registry. Therefore, it could not be found when importing the module
again.

This broke when we transitioned away from storing JavaScript modules in
a separate data structure and instead started treating them like any
other compilation units.

In order to fix this, generate synthetic QQmlType and QMetaType
instances also for JavaScript compilation units and store them in the
relevant data structures. This makes it possible to locate them when we
need them.

Pick-to: 6.9 6.8
Fixes: QTBUG-133053
Change-Id: I0425cd70a3745e793fec03105f58b708726ede06
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>