qmlformat: Have the outWriter ensure spaces instead of adding them
When formatting comments we try to preserve the spacing that the user
used around them. We keep track of how many spaces come before and after
a comment and insert them back accordingly.

However, when we continue to write out the following tokens we add
spaces to separate them. This space was being inserted unconditionally.
This could lead to a space being accounted for twice for certain comment
configurations. Once by the recorded amount of spaces after the comment
and once by the standard space to separate tokens. This would lead to
spaces being inserted in the formatted code that weren't there
originally in the user's code. One space would be inserted like this for
every format call.

To solve this, don't unconditionally insert a space to separate the
tokens but only make sure there is a space.

This issue can be reproduced in many "places" of the grammar. This patch
eliminates a good amount of them but not all.

Created QTBUG-133315 and QTBUG-133316.

Fixes: QTBUG-131386
Pick-to: 6.8
Change-Id: I25ac67f49a0198047912ef3f04dda3afc5cb9f66
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 711b643d5cff068794879698b2fe6ae6a4be8c2c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>