qmlformat: Accept `from` as an identifier
From is defined as an identifier with a special meaning in the context
of imports. This means it should generally be accepted as an identifier
in other contexts.

This fixes weird line indentation issues with the IndentingLineWriter.

Fixes: QTBUG-132065
Pick-to: 6.8
Change-Id: Iea5bf3af5d43af9a97f0f8e7867d12337e4041ca
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 5598bd16505f7ff6c8b0fd2cb48df8054d25782c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>