qqmljsutils: improve didYouMean performance
didYouMean eats up a lot of time in the profiler when imports can't be
found, so its not really a "performance fix". But it makes qmllint
faster on files that have a typo in its imports for example.

Swapping QList with QVarLengthArray makes the runtime of
tst_qmllint_benchmark:onlyQdsLintPlugin on
qtdesign-studio/examples/DesignEffectsDemo/content/Gallery.ui.qml go
from 150ms to 125ms when the QtDesign studio modules import paths are
missing. It seems that QVarLengthArray does less allocations than QList,
probably because of the std::swap.

It seemed that std::vector had similar performance to QVarLengthArray,
up to some ms.

Task-number: QTBUG-133349
Change-Id: I3facc8f194da6eaf1aebbce91292bf9b164b740e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>