Overlay: ignore right button clicks
So that ContextMenu can open when e.g. a Drawer is in the scene.

[ChangeLog][Important Behavior Changes][Controls][Overlay] The pressed
and released signals are no longer emitted for right clicks. This is to
allow ContextMenu to work when controls like Drawer are used.

Fixes: QTBUG-132765
Change-Id: If329a2ddaaf76e498e7ae2cd48a899fc23dba6b9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit fdb1db77a51e22857bda87dd0c6a783ea71d26f9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>