qmllint: avoid wasteful didYouMean calls
Add a new disabled mode to qqmljslogger that is used when warnings
emitted by qqmljsimportvisitor are actually not shown to the user. Add a
comment explaining the difference to a silenced logger (that still
processes warnings and can be used to generate a JSON representation of
the warnings) and a disabled logger (that does not contain any warning).

Disable the logger in QQmlJSTypeReader that is only used for the (lazy
file creation.

This avoids expensive didYouMean() calls that slow down the linter and
only create warnings that are ignored by the linter.

Also add some ifs to qqmljslogger to avoid any useless computations when
the logger is disabled anyway.

This commit makes the runtime of tst_qmllint_benchmark:onlyQdsLintPlugin
on qtdesign-studio/examples/DesignEffectsDemo/content/Gallery.ui.qml go
from 500ms to 135ms (mean over 3 iterations) when the QtDesign studio
modules import paths are unavailable.

Task-number: QTBUG-133349
Change-Id: I5abaa5e6776e39f2dee7f1a92577a240dd5c61c1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>