tst_qmllint_benchmark: also time "real" projects
Add an option to benchmark files from "real" QDS projects. The QDS
projects do not exist in the qtdeclarative repository, so the paths to
the demo repos should be set via environment variable. Also, the QDS
specific QML Modules only exist in the QDS installation path, so add an
environment variable for this too. Finally, some extra paths are needed
for the examples to run, so add a third environment variable to enable
the heuristics that adds probable import paths needed by the examples.

Add a check to make sure the import paths were passed correctly: there
should not be any 'Failed to import' messages, otherwise we end up
benchmarking the wrong thing.

Add a script that prints the environment for the benchmarks, so that you
can copy paste the output into Qt Creator's run environment widget.

Task-number: QTBUG-133349
Change-Id: I974fbeb5d87ff36bd5eea0b5c83d7fae73d00ec3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>