Menu: don't trigger a menu item on release if it's not highlighted
When a mouse press opens a menu, the press event does not highlight any
menu item (even if a menu item happens to be under the mouse cursor).
In that case, it takes an additional mouse move to highlight an item.

Behavior is now more consistent with appearance: it would be surprising
to activate a menu item on release if it was not highlighted already.
That was the worst aspect of QTBUG-133302, although it could have been
fixed in other ways (require the mouse to move past the drag threshold
as in qtbase dc1d3abd7e852829b9bc81c1ba9565f40147402d, or just move the
menu away from the cursor to prevent single-click accidents).

Ensure that QQuickAbstractButtonPrivate::animateTimer is both stopped
and reset to 0, so that it's testable. If animateClick() were called,
it would start the timer.

Also add logging of the release event to qt.quick.controls.popup.window
for easier troubleshooting.

Fixes: QTBUG-133302
Change-Id: I4c1cc8ec091e817c49456b4bc5ea1069566bce85
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit dac8529dc3fac2ab57be9c72ef7f436e94317e25)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>