DomEnvironment: stop using Structured Bindings
The moduleIndexWithUriHelper() function does everything right and
returns a proper struct with proper names, and not just pair or
tuple. So the caller ought not to have used SB to override those names
with their own ones.

There are many gotchas with SB that make it only acceptable as a
work-around for bad APIs, and this is not bad API.

The reason for doing the change now is that soon-to-be-released GCC-15
will (incorrectly) warn about capturing an SB value if said value is
only used as the initializer of an identically-named capture
variable. In this case:

  qqmldomtop.cpp:1547:65: error: captured structured bindings are a C++20 extension [-Werror=c++20-extensions]
   1547 |     std::shared_ptr<ModuleIndex> newModulePtr = [&, candidate = candidate](){
        |                                                                 ^~~~~~~~~
  qqmldomtop.cpp:1536:11: note: declared here
   1536 |     auto [candidate, origin] = moduleIndexWithUriHelper(self, uri, majorVersion, options);
        |           ^~~~~~~~~

Fix by receiving the result struct itself as `candidate` and using
their subobjects directly as opposed to aliasing them with SB.

This revealed two missing std::move()s, which this patch adds as a
drive-by (you cannot NRVO or automatically move from an SB variable).

As another drive-by, remove the unneeded extra empty () on the lambda
function.

Amends d2a2a5ec5868de1dfb7370d94a94c1e2766966c0.

Fixes: QTBUG-133323
Pick-to: 6.8 6.5
Change-Id: Ic6e1f3f68d12c3382d62746e9f960d6336ad2a40
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Liang Qi <liang.qi@qt.io>
(cherry picked from commit b7e341ec5a41dbc3a30880cd4e9a5d3d944813c8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>