Don't synthesize QContextMenuEvent if right-button opens a Menu
We don't ever want to see a single right mouse button event triggering
multiple context menus in different layers of a UI. It seems likely
that some previously-written QML applications already use TapHandlers to
open Menus; so the default context menus we are now adding to various
Controls could be redundant in some cases.

QQuickWindowPrivate::maybeSynthesizeContextMenuEvent() returns early if
QQuickWindowPrivate::rmbContextMenuEventEnabled == false;
QQuickMenu::itemChange() sets it false when it becomes visible.

So, adding a TapHandler and opening a custom menu is now a brute-force
way to prevent the built-in menu from opening. But merely customizing
the built-in menu should be possible in a cheaper way, so let's not
recommend doing it this way.

Change-Id: I40b071c82b03efbf36db7a840ae22363d1ec1585
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 555a2ed806bcda8c503b0e710703ad0f7bb27fd5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>