Avoid an allocation in ReferenceObject::init
We recently introduced an optimization to reduce the amount of reads
that we perform when dealing with QObject-provening properties.

As part of the optimization, we need to connect to the destroyed signal
of the original QObject, to ensure that we do perform a read to
invalidate our view of the data after the object is destroyed.

The receiver of the connection needs to modify the state of a
ReferenceObject that refers to the original QObject.

It is possible for the ReferenceObject to be destroyed before the
QObject is destroyed, so that we want to ensure that the connection we
created is dropped to avoid having the destroyed signal trigger a
callback that tries to mutate the state of memory that was freed.

To do so we need to hold an handle to the connection we introduce from
the ReferenceObject.

Due to the ReferenceObject object living in V4's managed memory, we need
it to be trivial.
As a `QMetaObject::Connection` handle is not trivial, we thus need to
add one layer of indirection by storing a pointer to it.
This means, in turn, that we are forced to perform an allocation for a
`QMetaObject::Connection` that will then copy from the original handle
we obtain.

To avoid the allocation, we instead store the connection in block
memory.

Thus, the handling of the connection that is introduced on an instance
of `ReferenceObject` was modified to store the handle in block memory.

The `onDelete` member, which stores the handle to the connection, of the
heap part of a `ReferenceObject` was modified a correctly aligned,
correctly sized byte array.

The code that takes care of producing the connection was modified to
avoid the allocation, instead directly allocating in the byte array by
placement new.

The code that takes care of disconnecting as necessary, in
`ReferenceObject::destroy`, was adapted to disconnect and destroy the
connection stored in the byte array.

Change-Id: I5472096cb163000545d7f632302df0182511f291
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>