tst_qquickTextEdit::preeditCursorRectangle verify edit has valid child
tst_qquicktextedit::preeditCursorRectangle() is flaky where in some
cases the child of edit is not found and a nullptr is returned. It was
verified if the child was a valid one, only after assigning it.

QTry_Verify if edit->child is a nullptr before assigning it.

The test will fail if no child is found for edit->findChild(). The
verificiation of the validity of the child is moved one line up and
made into a QTRY_VERIFY, as it is sometimes reached without processes
finished, creating a race condition.

Pick-to: 6.8 6.5
Change-Id: I1e83d6c3c4a29d1ac2046511a2943055f180d1b4
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit a61eb78f40eabd38346051da3cf071b03d069572)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>