tst_qquickmenu::popup: Make test less flaky
This test is very flaky when running it locally with the Material style.

Part of this has to do with the popup's x position sometimes being half
a pixel away from the intended location. Fix this by tolerating this
0.5 pixel difference. The reason behind this difference is currently
unknown.

Another issue is that we don't wait for the menu to close before
attempting to open it again.

Pick-to: 6.9 6.8
Fixes: QTBUG-133566
Change-Id: I08e68a87d477e6b3a9ea0b5b3350e5c8dc57919b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>