Set allocated memory to 0 on VxWorks
When VxWorks is configured with RTP_MEM_FILL false for performance
reasons, then it is not filling the allocated memory to 0, causing
the QML engine to crash, because it is expecting that it is filled
with 0s. This change handles this scenario by calling memset 0 on
the allocated memory. Qt needs to be configured with
QT_RTP_MEM_FILL define.

Task-number: QTBUG-133305
Pick-to: 6.8 6.9
Change-Id: I885c3cad7d81a11b8a0c125ddc098751b9643392
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>