Don't try to make layer textures larger than max size
If you try to allocate a layer which is larger than the maximum
texture size, we will get a warning from the graphics driver
which is not very informative.

While we cannot support layers larger than the max size, we can
catch the problem earlier and provide a better warning.

Pick-to: 6.5 6.8 6.9
Task-number: QTBUG-130879
Change-Id: If8998b5a316b07ec81b60619a33b468232f2deb7
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>