DialogsQuickImpl: add revision to signal and make properties FINAL
The sidebar is intended to be an implementation detail, and thus not
intended to be used by users directly. However, it makes sense for it to
have its properties flagged as FINAL for the performance gain. In
addition, a new signal didn't have a revision.

Found in API review.

Change-Id: I82b4314a3cf7f8336efe728a6c8da5d22a52a800
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit f024a3389e10118a9e9ae70f950b80d287cdd55b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>